BottomNavyBar está comum para as telas.
