+++ lib/pages/paginaCarrinho.dart
@@ -1,30 +1,32 @@
 import 'package:flutter/material.dart';
-class Carrinho extends StatefulWidget {
-  @override
-  _CarrinhoState createState() => _CarrinhoState();
-}
-
-class _CarrinhoState extends State<Carrinho> {
+import 'package:michellemirandastore/Materiais/BottomNavyBar.dart';
+class Carrinho extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text("Carrinho"),
-      ),
-    body: SafeArea(
-      minimum: EdgeInsets.fromLTRB(0, 0, 0, 15),
-        child: Column(
-          children: [
-            Container(
-              width: MediaQuery.of(context).size.width,
-//              height: MediaQuery.of(context).size.height,
-              child: Text(
-                ''
-              ),
-            )
-          ],
+        title: Text(
+          "Titulo"
         ),
       ),
+      body: Container(
+        constraints: BoxConstraints.expand(),
+            color: Colors.red,
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.stretch,
+              children: [
+                Container(
+                  height: 30,
+                  color: Colors.amber,
+                  child: Text(
+                    "-"
+                  ),
+                )
+              ],
+            )
+          ),
+      bottomNavigationBar: BottomNavyBar(),
     );
   }
 }
+
